chore(har-remix): eslint migration
